Requested PR changes

+++ include/pl/camera.h
@@ -3,7 +3,7 @@
 
 #include "common.h"
 
-void _plcamera_entrypoint_12(s32, u8*, u8*);
-s32 _plcamera_entrypoint_14(s32, u8*, u8*);
+void _plcamera_entrypoint_12(s32, f32*, f32*);
+s32 _plcamera_entrypoint_14(s32, f32*, f32*);
 
 #endif
